Refactoring (create package and move annotation); fix typo
