+++ dspace-spring-rest/src/main/java/org/dspace/app/rest/model/hateoas/BitstreamResource.java
@@ -8,6 +8,7 @@
 package org.dspace.app.rest.model.hateoas;
 
 import org.dspace.app.rest.model.BitstreamRest;
+import org.dspace.app.rest.model.hateoas.annotations.RelNameDSpaceResource;
 import org.dspace.app.rest.utils.Utils;
 
 /**
