+++ dspace-spring-rest/src/main/java/org/dspace/app/rest/model/hateoas/CommunityResource.java
@@ -15,6 +15,7 @@
 package org.dspace.app.rest.model.hateoas;
 
 import org.dspace.app.rest.model.CommunityRest;
+import org.dspace.app.rest.model.hateoas.annotations.RelNameDSpaceResource;
 import org.dspace.app.rest.utils.Utils;
 
 /**
