+++ dspace-spring-rest/src/main/java/org/dspace/app/rest/model/hateoas/DSpaceRelProvider.java
@@ -7,6 +7,7 @@
  */
 package org.dspace.app.rest.model.hateoas;
 
+import org.dspace.app.rest.model.hateoas.annotations.RelNameDSpaceResource;
 import org.springframework.hateoas.core.EvoInflectorRelProvider;
 
 /**
