+++ dspace-spring-rest/src/main/java/org/dspace/app/rest/repository/BrowseEntryLinkRepository.java
@@ -145,7 +145,7 @@ public class BrowseEntryLinkRepository extends AbstractDSpaceRestRepository
 	}
 	
 	@Override
-	public boolean isEmbbeddableRelation(Object data, String name) {
+	public boolean isEmbeddableRelation(Object data, String name) {
 		BrowseIndexRest bir = (BrowseIndexRest) data;
 		if (bir.isMetadataBrowse() && "entries".equals(name)) {
 			return true;
