+++ dspace-spring-rest/src/main/java/org/dspace/app/rest/repository/LinkRestRepository.java
@@ -20,7 +20,7 @@ import org.springframework.hateoas.ResourceSupport;
 public interface LinkRestRepository<L extends Serializable> {
 	public abstract ResourceSupport wrapResource(L model, String... rels);
 
-	public default boolean isEmbbeddableRelation(Object data, String name) {
+	public default boolean isEmbeddableRelation(Object data, String name) {
 		return true;
 	}
 }
